Refactor Filters step to make it more modular
- Added `FieldFiltersCard` component to display and manage filters for each field.
- Introduced `FilterInputs` for various input types including range, date, multi-select, and boolean.
- Created `FilterRow` to handle individual filter display and interactions.
- Developed `useFiltersManager` hook for managing filter state and logic.
- Implemented utility functions for filter schema retrieval, validation, and default value handling.
- Integrated filters into the main `StepFilters` component for user configuration.
