+++ components/text-gen-builder/index.ts
@@ -2,6 +2,15 @@ export { default as TemplateBuilder } from '../text-gen-builder';
 export * from './types';
 export { default as StepBuilder } from './steps/builder';
 export { default as StepFilters } from './steps/filters';
-export { default as AvailableFieldsList } from './available-fields-list';
-export { default as TemplateFieldsList } from './template-fields-list';
-export { default as FieldItem } from './field-item';
+
+// Builder sub-components
+export { default as AvailableFieldsList } from './steps/builder/AvailableFieldsList';
+export { default as TemplateFieldsList } from './steps/builder/TemplateFieldsList';
+export { default as FieldItem } from './steps/builder/FieldItem';
+export { default as IconColorPicker } from './steps/builder/IconColorPicker';
+export { default as AddFieldModal } from './steps/builder/AddFieldModal';
+
+// Builder utilities and hooks
+export { useDragAndDrop } from './steps/builder/useDragAndDrop';
+export { useScrollFade, ScrollFadeGradients } from './steps/builder/useScrollFade';
+export * from './steps/builder/constants';
